[Bugfix] fix some bugs in windows port

+++ enemies.c
@@ -1,7 +1,5 @@
 #include "enemies.h"
 
-#include <math.h>
-
 #include <tari/datastructures.h>
 #include <tari/animation.h>
 #include <tari/texture.h>
@@ -112,7 +110,7 @@ static ScriptPosition loadSingleEnemyTypeAnimation(ScriptPosition pos, Animation
 	animation->mDuration = v;
 	
 	int i;
-	for(i = 0; i < animation->mFrameAmount; i++) {
+	for(i = 0; i < (int)animation->mFrameAmount; i++) {
 		char file[100];
 		char path[100];
 		pos = getNextScriptString(pos, file);
