+++ player.c
@@ -97,7 +97,7 @@ static ScriptPosition loadTextureDataAndAnimation(ScriptPosition position, Textu
 	position = getNextScriptString(position, name);
 
 	int i;
-	for(i = 0; i < animation->mFrameAmount; i++) {
+	for(i = 0; i < (int)animation->mFrameAmount; i++) {
 		sprintf(path, "sprites/%s%d.pkg", name, i);
 		textureData[i] = loadTexture(path);
 	}
